refactor(renderer): split Renderer.ts into PrimitivePipeline and SpritePipeline
Renderer.ts was 1210 lines managing two entirely separate WebGPU
pipelines, exceeding the 800-line project maximum. Each pipeline had
its own WGSL shader, buffer setup, draw methods, and vertex batching
all mixed together.

Split along the natural pipeline boundary into three focused files:

- PrimitivePipeline.ts (~430 lines): colored geometry pipeline with
  its own WGSL shader, GPU buffers, and all primitive draw methods
  (pixel, line, rect, text). Exposes encodePass() and reset().

- SpritePipeline.ts (~430 lines): textured quad pipeline with its own
  WGSL shader, nearest-neighbor sampler, texture bind group cache, and
  batch flushing. Exposes encodePass() and reset().

- Renderer.ts (~265 lines): thin orchestrator that instantiates both
  pipelines, manages frame lifecycle (beginFrame/endFrame), owns the
  camera offset, and delegates all draw calls with one-liner methods.

No external API changes. BTAPI.ts and BlitTech.ts are untouched.
