+++ cspell.json
@@ -44,11 +44,12 @@
     "subsystem",
     "subsystems",
     "timestep",
+    "unflushed",
     "unorm",
+    "Václav",
     "Vančura",
     "Vite",
     "Vitest",
-    "Václav",
     "WebGPU",
     "WGSL",
     "xadvance",
